==> Code/6_EWS-ByPop.Rmd <==
# Assessing EWS by Population Grouping

Here, I'm just playing around with calculating AUCs by population, instead of as a composite.

# Required Packages

```{r, include = FALSE}
require(scales)
require(evobiR)
require(gdata)
require(doBy)
require(ecp)
require(strucchange)
require(TTR)
require(moments)
require(tidyverse)
```

# SUB-GROUPING SIMULATED DATA

We did this to look at variation!

# Raw, Sliding Window = 15 Days

```{r}
## Reading in Data
constant1 <- read.csv("Data/Simulated/constant.incidence.last.120.1.csv")
constant2 <- read.csv("Data/Simulated/constant.incidence.last.120.2.csv")
constant3 <- read.csv("Data/Simulated/constant.incidence.last.120.3.csv")
constant4 <- read.csv("Data/Simulated/constant.incidence.last.120.4.csv")
constant5 <- read.csv("Data/Simulated/constant.incidence.last.120.5.csv")
constant6 <- read.csv("Data/Simulated/constant.incidence.last.120.6.csv")
constant7 <- read.csv("Data/Simulated/constant.incidence.last.120.7.csv")
constant8 <- read.csv("Data/Simulated/constant.incidence.last.120.8.csv")
constant9 <- read.csv("Data/Simulated/constant.incidence.last.120.9.csv")
constant10 <- read.csv("Data/Simulated/constant.incidence.last.120.10.csv")
warming1 <- read.csv("Data/Simulated/warming.incidence.last.120.1.csv")
warming2 <- read.csv("Data/Simulated/warming.incidence.last.120.2.csv")
warming3 <- read.csv("Data/Simulated/warming.incidence.last.120.3.csv")
warming4 <- read.csv("Data/Simulated/warming.incidence.last.120.4.csv")
warming5 <- read.csv("Data/Simulated/warming.incidence.last.120.5.csv")
warming6 <- read.csv("Data/Simulated/warming.incidence.last.120.6.csv")
warming7 <- read.csv("Data/Simulated/warming.incidence.last.120.7.csv")
warming8 <- read.csv("Data/Simulated/warming.incidence.last.120.8.csv")
warming9 <- read.csv("Data/Simulated/warming.incidence.last.120.9.csv")
warming10 <- read.csv("Data/Simulated/warming.incidence.last.120.10.csv")

## Daily Incidence
constant.incidence.daily <- cbind(constant1[ ,2:101], constant2[ ,2:101], constant3[ ,2:101],
                                  constant4[ ,2:101], constant5[ ,2:101], constant6[ ,2:101],
                                  constant7[ ,2:101], constant8[ ,2:101], constant9[ ,2:101],
                                  constant10[ ,2:101])
warming.incidence.daily <- cbind(warming1[ ,2:101], warming2[ ,2:101], warming3[ ,2:101],
                                  warming4[ ,2:101], warming5[ ,2:101], warming6[ ,2:101],
                                  warming7[ ,2:101], warming8[ ,2:101], warming9[ ,2:101],
                                  warming10[ ,2:101])

## Standard Deviation, Mean, Skewness, Kurtosis
constant.sims.sd.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.mean.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.sk.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.kt.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.sd.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.mean.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.sk.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.kt.15 <- matrix(nrow = 105, ncol = num.of.sims)
for(i in 1:num.of.sims){
  constant.sims.sd.15[ ,i] <- SlidingWindow(FUN = "sd", data = constant.incidence.daily[,i], window = 15,step = 1)
  constant.sims.mean.15[ ,i] <- SlidingWindow(FUN = "mean", data = constant.incidence.daily[,i], window = 15, step = 1)
  constant.sims.sk.15[ ,i] <- SlidingWindow(FUN = "skewness", data = constant.incidence.daily[ ,i], window = 15, step = 1)
  constant.sims.kt.15[ ,i] <- SlidingWindow(FUN = "kurtosis", data = constant.incidence.daily[ ,i], window = 15, step = 1)
}
for(i in 1:num.of.sims){
  warming.sims.sd.15[ ,i] <- SlidingWindow(FUN = "sd", data = warming.incidence.daily[,i], window = 15,step = 1)
  warming.sims.mean.15[ ,i] <- SlidingWindow(FUN = "mean", data = warming.incidence.daily[,i], window = 15, step = 1)
  warming.sims.sk.15[ ,i] <- SlidingWindow(FUN = "skewness", data = warming.incidence.daily[ ,i], window = 15, step = 1)
  warming.sims.kt.15[ ,i] <- SlidingWindow(FUN = "kurtosis", data = warming.incidence.daily[ ,i], window = 15, step = 1)
}

## Variance
constant.sims.var.15 <- constant.sims.sd.15*constant.sims.sd.15
warming.sims.var.15 <- warming.sims.sd.15*warming.sims.sd.15

## Co-Efficient of Variation
constant.sims.cv.15 <- constant.sims.sd.15/constant.sims.mean.15
warming.sims.cv.15 <- warming.sims.sd.15/warming.sims.mean.15

## Index of Dispersion
constant.sims.id.15 <- constant.sims.var.15/constant.sims.mean.15
warming.sims.id.15 <- warming.sims.var.15/warming.sims.mean.15

## First Difference Variance
constant.sims.fdv.15 <- matrix(ncol = num.of.sims, nrow = 105)
warming.sims.fdv.15 <- matrix(ncol = num.of.sims, nrow = 105)
for(i in 1:num.of.sims){
  for(j in 2:105){
    constant.sims.fdv.15[j - 1,i] <- constant.sims.var.15[j,i] - constant.sims.var.15[j - 1,i]
    warming.sims.fdv.15[j - 1,i] <- warming.sims.var.15[j,i] - warming.sims.var.15[j - 1,i]
  }
}

## Autocorrelation and Autocovariance
constant.sims.ac.15 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.15 <- matrix(ncol = num.of.sims, nrow = 120)
for(j in 1:num.of.sims){
  for(i in 1:105){
    window.start <- i
    window.end <- i + 14
    segment <- constant.incidence.daily[c(window.start:window.end),j]
    constant.sims.ac.15[i, j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.15[i, j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
warming.sims.ac.15 <- matrix(ncol = num.of.sims,nrow = 120)
warming.sims.acv.15 <- matrix(ncol = num.of.sims,nrow = 120)
for(j in 1:num.of.sims){
  for(i in 1:105){
    window.start <- i
    window.end <- i + 14
    segment <- warming.incidence.daily[c(window.start:window.end),j]
    warming.sims.ac.15[i, j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.15[i, j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

## Decay Time
constant.sims.dt.15 <- matrix(ncol = num.of.sims, nrow = 45)
warming.sims.dt.15 <- matrix(nco = num.of.sims, nrow = 45)
for(i in 1:num.of.sims){
  for(j in 1:45){
    suppressWarnings(constant.sims.dt.15[j, i] <- -j/log(constant.sims.ac.15[j, i]))
    suppressWarnings(warming.sims.dt.15[j, i] <- -j/log(warming.sims.ac.15[j, i]))
  }
}

## Kendall's Tau
constant.sims.var.15 <- constant.sims.var.15[1:45, ]
constant.sims.mean.15 <- constant.sims.mean.15[1:45, ]
constant.sims.cv.15 <- constant.sims.cv.15[1:45, ]
constant.sims.sk.15 <- constant.sims.sk.15[1:45, ]
constant.sims.kt.15 <- constant.sims.kt.15[1:45, ]
constant.sims.id.15 <- constant.sims.id.15[1:45, ]
constant.sims.ac.15 <- constant.sims.ac.15[1:45, ]
constant.sims.acv.15 <- constant.sims.acv.15[1:45, ]
constant.sims.fdv.15 <- constant.sims.fdv.15[1:45, ]
constant.sims.dt.15 <- constant.sims.dt.15[1:45, ]
warming.sims.var.15 <- warming.sims.var.15[1:45, ]
warming.sims.mean.15 <- warming.sims.mean.15[1:45, ]
warming.sims.cv.15 <- warming.sims.cv.15[1:45, ]
warming.sims.sk.15 <- warming.sims.sk.15[1:45, ]
warming.sims.kt.15 <- warming.sims.kt.15[1:45, ]
warming.sims.id.15 <- warming.sims.id.15[1:45, ]
warming.sims.ac.15 <- warming.sims.ac.15[1:45, ]
warming.sims.acv.15 <- warming.sims.acv.15[1:45, ]
warming.sims.fdv.15 <-  warming.sims.fdv.15[1:45, ]
warming.sims.dt.15 <-  warming.sims.dt.15[1:45, ]
constant.sims.var.15.tau <- c()
constant.sims.mean.15.tau <- c()
constant.sims.cv.15.tau <- c()
constant.sims.sk.15.tau <- c()
constant.sims.kt.15.tau <- c()
constant.sims.id.15.tau <- c()
constant.sims.ac.15.tau <- c()
constant.sims.acv.15.tau <- c()
constant.sims.fdv.15.tau <- c()
constant.sims.dt.15.tau <- c()
warming.sims.var.15.tau <- c()
warming.sims.mean.15.tau <- c()
warming.sims.cv.15.tau <- c()
warming.sims.sk.15.tau <- c()
warming.sims.kt.15.tau <- c()
warming.sims.id.15.tau <- c()
warming.sims.ac.15.tau <- c()
warming.sims.acv.15.tau <- c()
warming.sims.fdv.15.tau <- c()
warming.sims.dt.15.tau <- c()
time.15 <- c(1:45)
for(i in 1:num.of.sims){
  constant.sims.mean.15.tau[i] <- cor(constant.sims.mean.15[ , i], time.15, method = "kendall")
  constant.sims.sk.15.tau[i] <- cor(constant.sims.sk.15[ , i], time.15, method = "kendall")
  constant.sims.kt.15.tau[i] <- cor(constant.sims.kt.15[ , i], time.15, method = "kendall")
  constant.sims.var.15.tau[i] <- cor(constant.sims.var.15[ , i], time.15, method = "kendall")
  constant.sims.cv.15.tau[i] <- cor(constant.sims.cv.15[ , i], time.15, method = "kendall")
  constant.sims.id.15.tau[i] <- cor(constant.sims.id.15[ , i], time.15, method = "kendall")
  constant.sims.fdv.15.tau[i] <- cor(constant.sims.fdv.15[ , i], time.15, method = "kendall")
  constant.sims.ac.15.tau[i] <- cor(constant.sims.ac.15[ , i], time.15, method = "kendall")
  constant.sims.acv.15.tau[i] <- cor(constant.sims.acv.15[ , i], time.15, method = "kendall")
  constant.sims.dt.15.tau[i] <- cor(constant.sims.dt.15[ , i], time.15, method = "kendall")
  
  warming.sims.mean.15.tau[i] <- cor(warming.sims.mean.15[ , i], time.15, method = "kendall")
  warming.sims.sk.15.tau[i] <- cor(warming.sims.sk.15[ , i], time.15, method = "kendall")
  warming.sims.kt.15.tau[i] <- cor(warming.sims.kt.15[ , i], time.15, method = "kendall")
  warming.sims.var.15.tau[i] <- cor(warming.sims.var.15[ , i], time.15, method = "kendall")
  warming.sims.cv.15.tau[i] <- cor(warming.sims.cv.15[ , i], time.15, method = "kendall")
  warming.sims.id.15.tau[i] <- cor(warming.sims.id.15[ , i], time.15, method = "kendall")
  warming.sims.fdv.15.tau[i] <- cor(warming.sims.fdv.15[ , i], time.15, method = "kendall")
  warming.sims.ac.15.tau[i] <- cor(warming.sims.ac.15[ , i], time.15, method = "kendall")
  warming.sims.acv.15.tau[i] <- cor(warming.sims.acv.15[ , i], time.15, method = "kendall")
  warming.sims.dt.15.tau[i] <- cor(warming.sims.dt.15[ , i], time.15, method = "kendall")
}
```

## Calculating AUCs by Grouping

### Mean

```{r}
c.grp1 <- constant.sims.mean.15.tau[1:100]
c.grp2 <- constant.sims.mean.15.tau[101:200]
c.grp3 <- constant.sims.mean.15.tau[201:300]
c.grp4 <- constant.sims.mean.15.tau[301:400]
c.grp5 <- constant.sims.mean.15.tau[401:500]
c.grp6 <- constant.sims.mean.15.tau[501:600]
c.grp7 <- constant.sims.mean.15.tau[601:700]
c.grp8 <- constant.sims.mean.15.tau[701:800]
c.grp9 <- constant.sims.mean.15.tau[801:900]
c.grp10 <- constant.sims.mean.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.mean.15.tau[1:100]
w.grp2 <- warming.sims.mean.15.tau[101:200]
w.grp3 <- warming.sims.mean.15.tau[201:300]
w.grp4 <- warming.sims.mean.15.tau[301:400]
w.grp5 <- warming.sims.mean.15.tau[401:500]
w.grp6 <- warming.sims.mean.15.tau[501:600]
w.grp7 <- warming.sims.mean.15.tau[601:700]
w.grp8 <- warming.sims.mean.15.tau[701:800]
w.grp9 <- warming.sims.mean.15.tau[801:900]
w.grp10 <- warming.sims.mean.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

m.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  m.aucs[i] <- calc_auc(test, null)
}
```

### Skewness

```{r}
c.grp1 <- constant.sims.sk.15.tau[1:100]
c.grp2 <- constant.sims.sk.15.tau[101:200]
c.grp3 <- constant.sims.sk.15.tau[201:300]
c.grp4 <- constant.sims.sk.15.tau[301:400]
c.grp5 <- constant.sims.sk.15.tau[401:500]
c.grp6 <- constant.sims.sk.15.tau[501:600]
c.grp7 <- constant.sims.sk.15.tau[601:700]
c.grp8 <- constant.sims.sk.15.tau[701:800]
c.grp9 <- constant.sims.sk.15.tau[801:900]
c.grp10 <- constant.sims.sk.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.sk.15.tau[1:100]
w.grp2 <- warming.sims.sk.15.tau[101:200]
w.grp3 <- warming.sims.sk.15.tau[201:300]
w.grp4 <- warming.sims.sk.15.tau[301:400]
w.grp5 <- warming.sims.sk.15.tau[401:500]
w.grp6 <- warming.sims.sk.15.tau[501:600]
w.grp7 <- warming.sims.sk.15.tau[601:700]
w.grp8 <- warming.sims.sk.15.tau[701:800]
w.grp9 <- warming.sims.sk.15.tau[801:900]
w.grp10 <- warming.sims.sk.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

skew.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  skew.aucs[i] <- calc_auc(test, null)
}
```

### Kurtosis

```{r}
c.grp1 <- constant.sims.kt.15.tau[1:100]
c.grp2 <- constant.sims.kt.15.tau[101:200]
c.grp3 <- constant.sims.kt.15.tau[201:300]
c.grp4 <- constant.sims.kt.15.tau[301:400]
c.grp5 <- constant.sims.kt.15.tau[401:500]
c.grp6 <- constant.sims.kt.15.tau[501:600]
c.grp7 <- constant.sims.kt.15.tau[601:700]
c.grp8 <- constant.sims.kt.15.tau[701:800]
c.grp9 <- constant.sims.kt.15.tau[801:900]
c.grp10 <- constant.sims.kt.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.kt.15.tau[1:100]
w.grp2 <- warming.sims.kt.15.tau[101:200]
w.grp3 <- warming.sims.kt.15.tau[201:300]
w.grp4 <- warming.sims.kt.15.tau[301:400]
w.grp5 <- warming.sims.kt.15.tau[401:500]
w.grp6 <- warming.sims.kt.15.tau[501:600]
w.grp7 <- warming.sims.kt.15.tau[601:700]
w.grp8 <- warming.sims.kt.15.tau[701:800]
w.grp9 <- warming.sims.kt.15.tau[801:900]
w.grp10 <- warming.sims.kt.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

k.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  k.aucs[i] <- calc_auc(test, null)
}
```

### Variance

```{r}
c.grp1 <- constant.sims.var.15.tau[1:100]
c.grp2 <- constant.sims.var.15.tau[101:200]
c.grp3 <- constant.sims.var.15.tau[201:300]
c.grp4 <- constant.sims.var.15.tau[301:400]
c.grp5 <- constant.sims.var.15.tau[401:500]
c.grp6 <- constant.sims.var.15.tau[501:600]
c.grp7 <- constant.sims.var.15.tau[601:700]
c.grp8 <- constant.sims.var.15.tau[701:800]
c.grp9 <- constant.sims.var.15.tau[801:900]
c.grp10 <- constant.sims.var.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.var.15.tau[1:100]
w.grp2 <- warming.sims.var.15.tau[101:200]
w.grp3 <- warming.sims.var.15.tau[201:300]
w.grp4 <- warming.sims.var.15.tau[301:400]
w.grp5 <- warming.sims.var.15.tau[401:500]
w.grp6 <- warming.sims.var.15.tau[501:600]
w.grp7 <- warming.sims.var.15.tau[601:700]
w.grp8 <- warming.sims.var.15.tau[701:800]
w.grp9 <- warming.sims.var.15.tau[801:900]
w.grp10 <- warming.sims.var.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

var.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  var.aucs[i] <- calc_auc(test, null)
}
```

### Co-Efficient of Variation

```{r}
c.grp1 <- constant.sims.cv.15.tau[1:100]
c.grp2 <- constant.sims.cv.15.tau[101:200]
c.grp3 <- constant.sims.cv.15.tau[201:300]
c.grp4 <- constant.sims.cv.15.tau[301:400]
c.grp5 <- constant.sims.cv.15.tau[401:500]
c.grp6 <- constant.sims.cv.15.tau[501:600]
c.grp7 <- constant.sims.cv.15.tau[601:700]
c.grp8 <- constant.sims.cv.15.tau[701:800]
c.grp9 <- constant.sims.cv.15.tau[801:900]
c.grp10 <- constant.sims.cv.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.cv.15.tau[1:100]
w.grp2 <- warming.sims.cv.15.tau[101:200]
w.grp3 <- warming.sims.cv.15.tau[201:300]
w.grp4 <- warming.sims.cv.15.tau[301:400]
w.grp5 <- warming.sims.cv.15.tau[401:500]
w.grp6 <- warming.sims.cv.15.tau[501:600]
w.grp7 <- warming.sims.cv.15.tau[601:700]
w.grp8 <- warming.sims.cv.15.tau[701:800]
w.grp9 <- warming.sims.cv.15.tau[801:900]
w.grp10 <- warming.sims.cv.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

cv.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  cv.aucs[i] <- calc_auc(test, null)
}
```

### Index of Dispersion

```{r}
c.grp1 <- constant.sims.id.15.tau[1:100]
c.grp2 <- constant.sims.id.15.tau[101:200]
c.grp3 <- constant.sims.id.15.tau[201:300]
c.grp4 <- constant.sims.id.15.tau[301:400]
c.grp5 <- constant.sims.id.15.tau[401:500]
c.grp6 <- constant.sims.id.15.tau[501:600]
c.grp7 <- constant.sims.id.15.tau[601:700]
c.grp8 <- constant.sims.id.15.tau[701:800]
c.grp9 <- constant.sims.id.15.tau[801:900]
c.grp10 <- constant.sims.id.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.id.15.tau[1:100]
w.grp2 <- warming.sims.id.15.tau[101:200]
w.grp3 <- warming.sims.id.15.tau[201:300]
w.grp4 <- warming.sims.id.15.tau[301:400]
w.grp5 <- warming.sims.id.15.tau[401:500]
w.grp6 <- warming.sims.id.15.tau[501:600]
w.grp7 <- warming.sims.id.15.tau[601:700]
w.grp8 <- warming.sims.id.15.tau[701:800]
w.grp9 <- warming.sims.id.15.tau[801:900]
w.grp10 <- warming.sims.id.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

id.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  id.aucs[i] <- calc_auc(test, null)
}
```

### First Difference Variance

```{r}
c.grp1 <- constant.sims.fdv.15.tau[1:100]
c.grp2 <- constant.sims.fdv.15.tau[101:200]
c.grp3 <- constant.sims.fdv.15.tau[201:300]
c.grp4 <- constant.sims.fdv.15.tau[301:400]
c.grp5 <- constant.sims.fdv.15.tau[401:500]
c.grp6 <- constant.sims.fdv.15.tau[501:600]
c.grp7 <- constant.sims.fdv.15.tau[601:700]
c.grp8 <- constant.sims.fdv.15.tau[701:800]
c.grp9 <- constant.sims.fdv.15.tau[801:900]
c.grp10 <- constant.sims.fdv.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.fdv.15.tau[1:100]
w.grp2 <- warming.sims.fdv.15.tau[101:200]
w.grp3 <- warming.sims.fdv.15.tau[201:300]
w.grp4 <- warming.sims.fdv.15.tau[301:400]
w.grp5 <- warming.sims.fdv.15.tau[401:500]
w.grp6 <- warming.sims.fdv.15.tau[501:600]
w.grp7 <- warming.sims.fdv.15.tau[601:700]
w.grp8 <- warming.sims.fdv.15.tau[701:800]
w.grp9 <- warming.sims.fdv.15.tau[801:900]
w.grp10 <- warming.sims.fdv.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

fdv.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  fdv.aucs[i] <- calc_auc(test, null)
}
```

### Autocorrelation

```{r}
c.grp1 <- constant.sims.ac.15.tau[1:100]
c.grp2 <- constant.sims.ac.15.tau[101:200]
c.grp3 <- constant.sims.ac.15.tau[201:300]
c.grp4 <- constant.sims.ac.15.tau[301:400]
c.grp5 <- constant.sims.ac.15.tau[401:500]
c.grp6 <- constant.sims.ac.15.tau[501:600]
c.grp7 <- constant.sims.ac.15.tau[601:700]
c.grp8 <- constant.sims.ac.15.tau[701:800]
c.grp9 <- constant.sims.ac.15.tau[801:900]
c.grp10 <- constant.sims.ac.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.ac.15.tau[1:100]
w.grp2 <- warming.sims.ac.15.tau[101:200]
w.grp3 <- warming.sims.ac.15.tau[201:300]
w.grp4 <- warming.sims.ac.15.tau[301:400]
w.grp5 <- warming.sims.ac.15.tau[401:500]
w.grp6 <- warming.sims.ac.15.tau[501:600]
w.grp7 <- warming.sims.ac.15.tau[601:700]
w.grp8 <- warming.sims.ac.15.tau[701:800]
w.grp9 <- warming.sims.ac.15.tau[801:900]
w.grp10 <- warming.sims.ac.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

ac.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  ac.aucs[i] <- calc_auc(test, null)
}
```

### Autocovariance

```{r}
c.grp1 <- constant.sims.acv.15.tau[1:100]
c.grp2 <- constant.sims.acv.15.tau[101:200]
c.grp3 <- constant.sims.acv.15.tau[201:300]
c.grp4 <- constant.sims.acv.15.tau[301:400]
c.grp5 <- constant.sims.acv.15.tau[401:500]
c.grp6 <- constant.sims.acv.15.tau[501:600]
c.grp7 <- constant.sims.acv.15.tau[601:700]
c.grp8 <- constant.sims.acv.15.tau[701:800]
c.grp9 <- constant.sims.acv.15.tau[801:900]
c.grp10 <- constant.sims.acv.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.acv.15.tau[1:100]
w.grp2 <- warming.sims.acv.15.tau[101:200]
w.grp3 <- warming.sims.acv.15.tau[201:300]
w.grp4 <- warming.sims.acv.15.tau[301:400]
w.grp5 <- warming.sims.acv.15.tau[401:500]
w.grp6 <- warming.sims.acv.15.tau[501:600]
w.grp7 <- warming.sims.acv.15.tau[601:700]
w.grp8 <- warming.sims.acv.15.tau[701:800]
w.grp9 <- warming.sims.acv.15.tau[801:900]
w.grp10 <- warming.sims.acv.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

acv.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  acv.aucs[i] <- calc_auc(test, null)
}
```

### Decay Time

```{r}
c.grp1 <- constant.sims.dt.15.tau[1:100]
c.grp2 <- constant.sims.dt.15.tau[101:200]
c.grp3 <- constant.sims.dt.15.tau[201:300]
c.grp4 <- constant.sims.dt.15.tau[301:400]
c.grp5 <- constant.sims.dt.15.tau[401:500]
c.grp6 <- constant.sims.dt.15.tau[501:600]
c.grp7 <- constant.sims.dt.15.tau[601:700]
c.grp8 <- constant.sims.dt.15.tau[701:800]
c.grp9 <- constant.sims.dt.15.tau[801:900]
c.grp10 <- constant.sims.dt.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.dt.15.tau[1:100]
w.grp2 <- warming.sims.dt.15.tau[101:200]
w.grp3 <- warming.sims.dt.15.tau[201:300]
w.grp4 <- warming.sims.dt.15.tau[301:400]
w.grp5 <- warming.sims.dt.15.tau[401:500]
w.grp6 <- warming.sims.dt.15.tau[501:600]
w.grp7 <- warming.sims.dt.15.tau[601:700]
w.grp8 <- warming.sims.dt.15.tau[701:800]
w.grp9 <- warming.sims.dt.15.tau[801:900]
w.grp10 <- warming.sims.dt.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

dt.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  dt.aucs[i] <- calc_auc(test, null)
}
```

## Collecting AUCs

```{r}
raw_aucs <- c(m.aucs, skew.aucs, k.aucs, var.aucs, cv.aucs, id.aucs, fdv.aucs, ac.aucs, acv.aucs, dt.aucs)
```

## Plotting

```{r}
metric <- c(rep(1, 10), rep(2, 10), rep(3, 10), rep(4, 10), rep(5, 10),
            rep(6, 10), rep(7, 10), rep(8, 10), rep(9, 10), rep(10, 10))
raw_df <- data.frame(cbind(metric, raw_aucs))

raw_df$metric <- as.factor(raw_df$metric)
ggplot(raw_df, aes(x = metric, y = raw_aucs, colour = metric)) + 
  geom_boxplot() +
  geom_jitter(shape = 16, position = position_jitter(0.2)) +
  geom_hline(yintercept = 0.5, colour = "grey", linetype = "dotted") +
  theme(panel.background = element_blank(),
        plot.title = element_text(size = 18),
        axis.title = element_text(size = 16),
        axis.text.x = element_text(color = "black", size = 12, angle = 30, vjust = 0.8, hjust = 0.8),
        axis.text.y = element_text(color = "black", size = 12),
        legend.position = "none") +
  labs(title = "Raw Simulated Data | 15 Day Sliding Window", x = "", y = "AUC") +
  scale_x_discrete(labels = metrics_full) +
  ylim(0, 1)
```

# Pre-Processed, Sliding Window = 15 (Bandwidth = 3)

```{r}
## Read in Data
smooth.sim.const.b3 <- readRDS("Objects/smooth-sim-const-b3.rds")
smooth.sim.warm.b3 <- readRDS("Objects/smooth-sim-warm-b3.rds")

## Standard Deviation, Mean, Skewness, Kurtosis
# Making empty matricies
constant.sims.sd.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.mean.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.sk.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.kt.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.sd.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.mean.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.sk.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.kt.15 <- matrix(nrow = 105, ncol = num.of.sims)
for(i in 1:num.of.sims){
  constant.sims.sd.15[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.const.b3[,i], window = 15,step = 1)
  constant.sims.mean.15[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.const.b3[,i], window = 15, step = 1)
  constant.sims.sk.15[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.const.b3[,i], window = 15, step = 1)
  constant.sims.kt.15[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.const.b3[,i], window = 15, step = 1)
}
for(i in 1:num.of.sims){
  warming.sims.sd.15[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.warm.b3[,i], window = 15,step = 1)
  warming.sims.mean.15[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.warm.b3[,i], window = 15, step = 1)
  warming.sims.sk.15[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.warm.b3[,i], window = 15, step = 1)
  warming.sims.kt.15[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.warm.b3[,i], window = 15, step = 1)
}

## Variance
constant.sims.var.15 <- constant.sims.sd.15*constant.sims.sd.15
warming.sims.var.15 <- warming.sims.sd.15*warming.sims.sd.15

## Co-Efficient of Variation
constant.sims.cv.15 <- constant.sims.sd.15/constant.sims.mean.15
warming.sims.cv.15 <- warming.sims.sd.15/warming.sims.mean.15

## Index of Dispersion
constant.sims.id.15 <- constant.sims.var.15/constant.sims.mean.15
warming.sims.id.15 <- warming.sims.var.15/warming.sims.mean.15

## First Difference Variance
constant.sims.fdv.15 <- matrix(ncol = num.of.sims, nrow = 105)
warming.sims.fdv.15 <- matrix(ncol = num.of.sims, nrow = 105)
for(i in 1:num.of.sims){
  for(j in 2:105){
    constant.sims.fdv.15[j - 1,i] <- constant.sims.var.15[j,i] - constant.sims.var.15[j - 1,i]
    warming.sims.fdv.15[j - 1,i] <- warming.sims.var.15[j,i] - warming.sims.var.15[j - 1,i]
  }
}
constant.sims.fdv.15 <- constant.sims.fdv.15[1:104, ]
warming.sims.fdv.15 <- warming.sims.fdv.15[1:104, ]

## Autocorrelation and Autocovariance
constant.sims.ac.15 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.15 <- matrix(ncol = num.of.sims, nrow = 120)
for(j in 1:num.of.sims){
  for(i in 1:105){
    window.start <- i
    window.end <- i + 14
    segment <- smooth.sim.const.b3[c(window.start:window.end),j]
    constant.sims.ac.15[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.15[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
warming.sims.ac.15 <- matrix(ncol=num.of.sims,nrow = 120)
warming.sims.acv.15 <- matrix(ncol=num.of.sims,nrow = 120)
for(j in 1:num.of.sims){
  for(i in 1:105){
    window.start <- i
    window.end <- i + 14
    segment <- smooth.sim.warm.b3[c(window.start:window.end),j]
    warming.sims.ac.15[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.15[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

## Decay Time
constant.sims.dt.15 <- matrix(ncol = num.of.sims, nrow = 45)
warming.sims.dt.15 <- matrix(nco = num.of.sims, nrow = 45)
for(i in 1:num.of.sims){
  for(j in 1:45){
    suppressWarnings(constant.sims.dt.15[j, i] <- -j/log(constant.sims.ac.15[j, i]))
    suppressWarnings(warming.sims.dt.15[j, i] <- -j/log(warming.sims.ac.15[j, i]))
  }
}

## Kendall's Tau
constant.sims.var.15 <- constant.sims.var.15[1:45, ]
constant.sims.mean.15 <- constant.sims.mean.15[1:45, ]
constant.sims.cv.15 <- constant.sims.cv.15[1:45, ]
constant.sims.sk.15 <-  constant.sims.sk.15[1:45, ]
constant.sims.kt.15 <-  constant.sims.kt.15[1:45, ]
constant.sims.id.15 <-  constant.sims.id.15[1:45, ]
constant.sims.ac.15 <-  constant.sims.ac.15[1:45, ]
constant.sims.acv.15 <-  constant.sims.acv.15[1:45, ]
constant.sims.fdv.15 <-  constant.sims.fdv.15[1:45, ]
constant.sims.dt.15 <-  constant.sims.dt.15[1:45, ]
warming.sims.var.15 <- warming.sims.var.15[1:45, ]
warming.sims.mean.15 <- warming.sims.mean.15[1:45, ]
warming.sims.cv.15 <- warming.sims.cv.15[1:45, ]
warming.sims.sk.15 <-  warming.sims.sk.15[1:45, ]
warming.sims.kt.15 <-  warming.sims.kt.15[1:45, ]
warming.sims.id.15 <-  warming.sims.id.15[1:45, ]
warming.sims.ac.15 <-  warming.sims.ac.15[1:45, ]
warming.sims.acv.15 <-  warming.sims.acv.15[1:45, ]
warming.sims.fdv.15 <-  warming.sims.fdv.15[1:45, ]
warming.sims.dt.15 <-  warming.sims.dt.15[1:45, ]
constant.sims.var.15.tau <- c()
constant.sims.mean.15.tau <- c()
constant.sims.cv.15.tau <- c()
constant.sims.sk.15.tau <-  c()
constant.sims.kt.15.tau <- c()
constant.sims.id.15.tau <- c()
constant.sims.ac.15.tau <- c()
constant.sims.acv.15.tau <- c()
constant.sims.fdv.15.tau <-  c()
constant.sims.dt.15.tau <- c()
warming.sims.var.15.tau <- c()
warming.sims.mean.15.tau <- c()
warming.sims.cv.15.tau <- c()
warming.sims.sk.15.tau <-  c()
warming.sims.kt.15.tau <- c()
warming.sims.id.15.tau <-  c()
warming.sims.ac.15.tau <-  c()
warming.sims.acv.15.tau <-  c()
warming.sims.fdv.15.tau <-  c()
warming.sims.dt.15.tau <-  c()
time.15 <- c(1:45)
for(i in 1:num.of.sims){
  constant.sims.mean.15.tau[i] <- cor(constant.sims.mean.15[ , i], time.15, method = "kendall")
  constant.sims.sk.15.tau[i] <- cor(constant.sims.sk.15[ , i], time.15, method = "kendall")
  constant.sims.kt.15.tau[i] <- cor(constant.sims.kt.15[ , i], time.15, method = "kendall")
  constant.sims.var.15.tau[i] <- cor(constant.sims.var.15[ , i], time.15, method = "kendall")
  constant.sims.cv.15.tau[i] <- cor(constant.sims.cv.15[ , i], time.15, method = "kendall")
  constant.sims.id.15.tau[i] <- cor(constant.sims.id.15[ , i], time.15, method = "kendall")
  constant.sims.fdv.15.tau[i] <- cor(constant.sims.fdv.15[ , i], time.15, method = "kendall")
  constant.sims.ac.15.tau[i] <- cor(constant.sims.ac.15[ , i], time.15, method = "kendall")
  constant.sims.acv.15.tau[i] <- cor(constant.sims.acv.15[ , i], time.15, method = "kendall")
  constant.sims.dt.15.tau[i] <- cor(constant.sims.dt.15[ , i], time.15, method = "kendall")
  
  warming.sims.mean.15.tau[i] <- cor(warming.sims.mean.15[ , i], time.15, method = "kendall")
  warming.sims.sk.15.tau[i] <- cor(warming.sims.sk.15[ , i], time.15, method = "kendall")
  warming.sims.kt.15.tau[i] <- cor(warming.sims.kt.15[ , i], time.15, method = "kendall")
  warming.sims.var.15.tau[i] <- cor(warming.sims.var.15[ , i], time.15, method = "kendall")
  warming.sims.cv.15.tau[i] <- cor(warming.sims.cv.15[ , i], time.15, method = "kendall")
  warming.sims.id.15.tau[i] <- cor(warming.sims.id.15[ , i], time.15, method = "kendall")
  warming.sims.fdv.15.tau[i] <- cor(warming.sims.fdv.15[ , i], time.15, method = "kendall")
  warming.sims.ac.15.tau[i] <- cor(warming.sims.ac.15[ , i], time.15, method = "kendall")
  warming.sims.acv.15.tau[i] <- cor(warming.sims.acv.15[ , i], time.15, method = "kendall")
  warming.sims.dt.15.tau[i] <- cor(warming.sims.dt.15[ , i], time.15, method = "kendall")
}
```

## Calculating AUCs by Grouping

```{r}
c.grp1 <- constant.sims.var.15.tau[1:100]
c.grp2 <- constant.sims.var.15.tau[101:200]
c.grp3 <- constant.sims.var.15.tau[201:300]
c.grp4 <- constant.sims.var.15.tau[301:400]
c.grp5 <- constant.sims.var.15.tau[401:500]
c.grp6 <- constant.sims.var.15.tau[501:600]
c.grp7 <- constant.sims.var.15.tau[601:700]
c.grp8 <- constant.sims.var.15.tau[701:800]
c.grp9 <- constant.sims.var.15.tau[801:900]
c.grp10 <- constant.sims.var.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.var.15.tau[1:100]
w.grp2 <- warming.sims.var.15.tau[101:200]
w.grp3 <- warming.sims.var.15.tau[201:300]
w.grp4 <- warming.sims.var.15.tau[301:400]
w.grp5 <- warming.sims.var.15.tau[401:500]
w.grp6 <- warming.sims.var.15.tau[501:600]
w.grp7 <- warming.sims.var.15.tau[601:700]
w.grp8 <- warming.sims.var.15.tau[701:800]
w.grp9 <- warming.sims.var.15.tau[801:900]
w.grp10 <- warming.sims.var.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

var.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  var.aucs[i] <- calc_auc(test, null)
}
```

### Mean

```{r}
c.grp1 <- constant.sims.mean.15.tau[1:100]
c.grp2 <- constant.sims.mean.15.tau[101:200]
c.grp3 <- constant.sims.mean.15.tau[201:300]
c.grp4 <- constant.sims.mean.15.tau[301:400]
c.grp5 <- constant.sims.mean.15.tau[401:500]
c.grp6 <- constant.sims.mean.15.tau[501:600]
c.grp7 <- constant.sims.mean.15.tau[601:700]
c.grp8 <- constant.sims.mean.15.tau[701:800]
c.grp9 <- constant.sims.mean.15.tau[801:900]
c.grp10 <- constant.sims.mean.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.mean.15.tau[1:100]
w.grp2 <- warming.sims.mean.15.tau[101:200]
w.grp3 <- warming.sims.mean.15.tau[201:300]
w.grp4 <- warming.sims.mean.15.tau[301:400]
w.grp5 <- warming.sims.mean.15.tau[401:500]
w.grp6 <- warming.sims.mean.15.tau[501:600]
w.grp7 <- warming.sims.mean.15.tau[601:700]
w.grp8 <- warming.sims.mean.15.tau[701:800]
w.grp9 <- warming.sims.mean.15.tau[801:900]
w.grp10 <- warming.sims.mean.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

m.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  m.aucs[i] <- calc_auc(test, null)
}
```

### Skewness

```{r}
c.grp1 <- constant.sims.sk.15.tau[1:100]
c.grp2 <- constant.sims.sk.15.tau[101:200]
c.grp3 <- constant.sims.sk.15.tau[201:300]
c.grp4 <- constant.sims.sk.15.tau[301:400]
c.grp5 <- constant.sims.sk.15.tau[401:500]
c.grp6 <- constant.sims.sk.15.tau[501:600]
c.grp7 <- constant.sims.sk.15.tau[601:700]
c.grp8 <- constant.sims.sk.15.tau[701:800]
c.grp9 <- constant.sims.sk.15.tau[801:900]
c.grp10 <- constant.sims.sk.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.sk.15.tau[1:100]
w.grp2 <- warming.sims.sk.15.tau[101:200]
w.grp3 <- warming.sims.sk.15.tau[201:300]
w.grp4 <- warming.sims.sk.15.tau[301:400]
w.grp5 <- warming.sims.sk.15.tau[401:500]
w.grp6 <- warming.sims.sk.15.tau[501:600]
w.grp7 <- warming.sims.sk.15.tau[601:700]
w.grp8 <- warming.sims.sk.15.tau[701:800]
w.grp9 <- warming.sims.sk.15.tau[801:900]
w.grp10 <- warming.sims.sk.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

skew.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  skew.aucs[i] <- calc_auc(test, null)
}
```

### Kurtosis

```{r}
c.grp1 <- constant.sims.kt.15.tau[1:100]
c.grp2 <- constant.sims.kt.15.tau[101:200]
c.grp3 <- constant.sims.kt.15.tau[201:300]
c.grp4 <- constant.sims.kt.15.tau[301:400]
c.grp5 <- constant.sims.kt.15.tau[401:500]
c.grp6 <- constant.sims.kt.15.tau[501:600]
c.grp7 <- constant.sims.kt.15.tau[601:700]
c.grp8 <- constant.sims.kt.15.tau[701:800]
c.grp9 <- constant.sims.kt.15.tau[801:900]
c.grp10 <- constant.sims.kt.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.kt.15.tau[1:100]
w.grp2 <- warming.sims.kt.15.tau[101:200]
w.grp3 <- warming.sims.kt.15.tau[201:300]
w.grp4 <- warming.sims.kt.15.tau[301:400]
w.grp5 <- warming.sims.kt.15.tau[401:500]
w.grp6 <- warming.sims.kt.15.tau[501:600]
w.grp7 <- warming.sims.kt.15.tau[601:700]
w.grp8 <- warming.sims.kt.15.tau[701:800]
w.grp9 <- warming.sims.kt.15.tau[801:900]
w.grp10 <- warming.sims.kt.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

k.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  k.aucs[i] <- calc_auc(test, null)
}
```

### Variance

```{r}
c.grp1 <- constant.sims.var.15.tau[1:100]
c.grp2 <- constant.sims.var.15.tau[101:200]
c.grp3 <- constant.sims.var.15.tau[201:300]
c.grp4 <- constant.sims.var.15.tau[301:400]
c.grp5 <- constant.sims.var.15.tau[401:500]
c.grp6 <- constant.sims.var.15.tau[501:600]
c.grp7 <- constant.sims.var.15.tau[601:700]
c.grp8 <- constant.sims.var.15.tau[701:800]
c.grp9 <- constant.sims.var.15.tau[801:900]
c.grp10 <- constant.sims.var.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.var.15.tau[1:100]
w.grp2 <- warming.sims.var.15.tau[101:200]
w.grp3 <- warming.sims.var.15.tau[201:300]
w.grp4 <- warming.sims.var.15.tau[301:400]
w.grp5 <- warming.sims.var.15.tau[401:500]
w.grp6 <- warming.sims.var.15.tau[501:600]
w.grp7 <- warming.sims.var.15.tau[601:700]
w.grp8 <- warming.sims.var.15.tau[701:800]
w.grp9 <- warming.sims.var.15.tau[801:900]
w.grp10 <- warming.sims.var.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

var.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  var.aucs[i] <- calc_auc(test, null)
}
```

### Co-Efficient of Variation

```{r}
c.grp1 <- constant.sims.cv.15.tau[1:100]
c.grp2 <- constant.sims.cv.15.tau[101:200]
c.grp3 <- constant.sims.cv.15.tau[201:300]
c.grp4 <- constant.sims.cv.15.tau[301:400]
c.grp5 <- constant.sims.cv.15.tau[401:500]
c.grp6 <- constant.sims.cv.15.tau[501:600]
c.grp7 <- constant.sims.cv.15.tau[601:700]
c.grp8 <- constant.sims.cv.15.tau[701:800]
c.grp9 <- constant.sims.cv.15.tau[801:900]
c.grp10 <- constant.sims.cv.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.cv.15.tau[1:100]
w.grp2 <- warming.sims.cv.15.tau[101:200]
w.grp3 <- warming.sims.cv.15.tau[201:300]
w.grp4 <- warming.sims.cv.15.tau[301:400]
w.grp5 <- warming.sims.cv.15.tau[401:500]
w.grp6 <- warming.sims.cv.15.tau[501:600]
w.grp7 <- warming.sims.cv.15.tau[601:700]
w.grp8 <- warming.sims.cv.15.tau[701:800]
w.grp9 <- warming.sims.cv.15.tau[801:900]
w.grp10 <- warming.sims.cv.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

cv.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  cv.aucs[i] <- calc_auc(test, null)
}
```

### Index of Dispersion

```{r}
c.grp1 <- constant.sims.id.15.tau[1:100]
c.grp2 <- constant.sims.id.15.tau[101:200]
c.grp3 <- constant.sims.id.15.tau[201:300]
c.grp4 <- constant.sims.id.15.tau[301:400]
c.grp5 <- constant.sims.id.15.tau[401:500]
c.grp6 <- constant.sims.id.15.tau[501:600]
c.grp7 <- constant.sims.id.15.tau[601:700]
c.grp8 <- constant.sims.id.15.tau[701:800]
c.grp9 <- constant.sims.id.15.tau[801:900]
c.grp10 <- constant.sims.id.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.id.15.tau[1:100]
w.grp2 <- warming.sims.id.15.tau[101:200]
w.grp3 <- warming.sims.id.15.tau[201:300]
w.grp4 <- warming.sims.id.15.tau[301:400]
w.grp5 <- warming.sims.id.15.tau[401:500]
w.grp6 <- warming.sims.id.15.tau[501:600]
w.grp7 <- warming.sims.id.15.tau[601:700]
w.grp8 <- warming.sims.id.15.tau[701:800]
w.grp9 <- warming.sims.id.15.tau[801:900]
w.grp10 <- warming.sims.id.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

id.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  id.aucs[i] <- calc_auc(test, null)
}
```

### First Difference Variance

```{r}
c.grp1 <- constant.sims.fdv.15.tau[1:100]
c.grp2 <- constant.sims.fdv.15.tau[101:200]
c.grp3 <- constant.sims.fdv.15.tau[201:300]
c.grp4 <- constant.sims.fdv.15.tau[301:400]
c.grp5 <- constant.sims.fdv.15.tau[401:500]
c.grp6 <- constant.sims.fdv.15.tau[501:600]
c.grp7 <- constant.sims.fdv.15.tau[601:700]
c.grp8 <- constant.sims.fdv.15.tau[701:800]
c.grp9 <- constant.sims.fdv.15.tau[801:900]
c.grp10 <- constant.sims.fdv.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.fdv.15.tau[1:100]
w.grp2 <- warming.sims.fdv.15.tau[101:200]
w.grp3 <- warming.sims.fdv.15.tau[201:300]
w.grp4 <- warming.sims.fdv.15.tau[301:400]
w.grp5 <- warming.sims.fdv.15.tau[401:500]
w.grp6 <- warming.sims.fdv.15.tau[501:600]
w.grp7 <- warming.sims.fdv.15.tau[601:700]
w.grp8 <- warming.sims.fdv.15.tau[701:800]
w.grp9 <- warming.sims.fdv.15.tau[801:900]
w.grp10 <- warming.sims.fdv.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

fdv.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  fdv.aucs[i] <- calc_auc(test, null)
}
```

### Autocorrelation

```{r}
c.grp1 <- constant.sims.ac.15.tau[1:100]
c.grp2 <- constant.sims.ac.15.tau[101:200]
c.grp3 <- constant.sims.ac.15.tau[201:300]
c.grp4 <- constant.sims.ac.15.tau[301:400]
c.grp5 <- constant.sims.ac.15.tau[401:500]
c.grp6 <- constant.sims.ac.15.tau[501:600]
c.grp7 <- constant.sims.ac.15.tau[601:700]
c.grp8 <- constant.sims.ac.15.tau[701:800]
c.grp9 <- constant.sims.ac.15.tau[801:900]
c.grp10 <- constant.sims.ac.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.ac.15.tau[1:100]
w.grp2 <- warming.sims.ac.15.tau[101:200]
w.grp3 <- warming.sims.ac.15.tau[201:300]
w.grp4 <- warming.sims.ac.15.tau[301:400]
w.grp5 <- warming.sims.ac.15.tau[401:500]
w.grp6 <- warming.sims.ac.15.tau[501:600]
w.grp7 <- warming.sims.ac.15.tau[601:700]
w.grp8 <- warming.sims.ac.15.tau[701:800]
w.grp9 <- warming.sims.ac.15.tau[801:900]
w.grp10 <- warming.sims.ac.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

ac.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  ac.aucs[i] <- calc_auc(test, null)
}
```

### Autocovariance

```{r}
c.grp1 <- constant.sims.acv.15.tau[1:100]
c.grp2 <- constant.sims.acv.15.tau[101:200]
c.grp3 <- constant.sims.acv.15.tau[201:300]
c.grp4 <- constant.sims.acv.15.tau[301:400]
c.grp5 <- constant.sims.acv.15.tau[401:500]
c.grp6 <- constant.sims.acv.15.tau[501:600]
c.grp7 <- constant.sims.acv.15.tau[601:700]
c.grp8 <- constant.sims.acv.15.tau[701:800]
c.grp9 <- constant.sims.acv.15.tau[801:900]
c.grp10 <- constant.sims.acv.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.acv.15.tau[1:100]
w.grp2 <- warming.sims.acv.15.tau[101:200]
w.grp3 <- warming.sims.acv.15.tau[201:300]
w.grp4 <- warming.sims.acv.15.tau[301:400]
w.grp5 <- warming.sims.acv.15.tau[401:500]
w.grp6 <- warming.sims.acv.15.tau[501:600]
w.grp7 <- warming.sims.acv.15.tau[601:700]
w.grp8 <- warming.sims.acv.15.tau[701:800]
w.grp9 <- warming.sims.acv.15.tau[801:900]
w.grp10 <- warming.sims.acv.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

acv.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  acv.aucs[i] <- calc_auc(test, null)
}
```

### Decay Time

```{r}
c.grp1 <- constant.sims.dt.15.tau[1:100]
c.grp2 <- constant.sims.dt.15.tau[101:200]
c.grp3 <- constant.sims.dt.15.tau[201:300]
c.grp4 <- constant.sims.dt.15.tau[301:400]
c.grp5 <- constant.sims.dt.15.tau[401:500]
c.grp6 <- constant.sims.dt.15.tau[501:600]
c.grp7 <- constant.sims.dt.15.tau[601:700]
c.grp8 <- constant.sims.dt.15.tau[701:800]
c.grp9 <- constant.sims.dt.15.tau[801:900]
c.grp10 <- constant.sims.dt.15.tau[901:1000]
const.taus <- cbind(c.grp1, c.grp2, c.grp3, c.grp4, c.grp5,
                    c.grp6, c.grp7, c.grp8, c.grp9, c.grp10)

w.grp1 <- warming.sims.dt.15.tau[1:100]
w.grp2 <- warming.sims.dt.15.tau[101:200]
w.grp3 <- warming.sims.dt.15.tau[201:300]
w.grp4 <- warming.sims.dt.15.tau[301:400]
w.grp5 <- warming.sims.dt.15.tau[401:500]
w.grp6 <- warming.sims.dt.15.tau[501:600]
w.grp7 <- warming.sims.dt.15.tau[601:700]
w.grp8 <- warming.sims.dt.15.tau[701:800]
w.grp9 <- warming.sims.dt.15.tau[801:900]
w.grp10 <- warming.sims.dt.15.tau[901:1000]
warm.taus <- cbind(w.grp1, w.grp2, w.grp3, w.grp4, w.grp5,
                   w.grp6, w.grp7, w.grp8, w.grp9, w.grp10)

dt.aucs <- c()
for (i in 1:10){
  test <- warm.taus[ ,i]; null <- const.taus[ ,i]
  dt.aucs[i] <- calc_auc(test, null)
}
```

## Collecting AUCs

```{r}
smooth_3_aucs <- c(m.aucs, skew.aucs, k.aucs, var.aucs, cv.aucs, id.aucs, fdv.aucs, ac.aucs, acv.aucs, dt.aucs)
```

## Plotting

```{r}
metric <- c(rep(1, 10), rep(2, 10), rep(3, 10), rep(4, 10), rep(5, 10),
            rep(6, 10), rep(7, 10), rep(8, 10), rep(9, 10), rep(10, 10))
smooth_df <- data.frame(cbind(metric, smooth_3_aucs))

smooth_df$metric <- as.factor(smooth_df$metric)
ggplot(smooth_df, aes(x = metric, y = smooth_3_aucs, colour = metric)) + 
  geom_boxplot() +
  geom_jitter(shape = 16, position = position_jitter(0.2)) +
  geom_hline(yintercept = 0.5, colour = "grey", linetype = "dotted") +
  theme(panel.background = element_blank(),
        plot.title = element_text(size = 18),
        axis.title = element_text(size = 16),
        axis.text.x = element_text(color = "black", size = 12, angle = 30, vjust = 0.8, hjust = 0.8),
        axis.text.y = element_text(color = "black", size = 12),
        legend.position = "none") +
  labs(title = "Pre-Processed Simulated Data (Bandwidth = 3) | 15 Day Sliding Window", x = "", y = "AUC") +
  labs(x = "", y = "AUC") +
  scale_x_discrete(labels = metrics_full) +
  ylim(0, 1)
```

# BY POPULATION

Okay, so from Harris et al. (2020), "Early warning signals of malaria resurgence in Kericho, Kenya":

-   To test this prediction, from December 1981 to April 1993 ("approach to criticality"), **we calculated each of these 10 leading indicators over a rolling window** (i.e. the statistics were repeatedly calculated within a fixed size temporal subsample of the data, shifting the window by one month for each computation). To better distinguish the signal of critical slowing down from background noise and periodic trends, we preprocessed the raw case data by detrending with a Gaussian kernel...

-   **All analyses were conducted in R using the spaero package**, a freely available package that provides a wide range of methods for investigating critical slowing down in time series. Data and code for reproducing analyses reported in this paper are available online at [https://github.com/mjharris95/Kericho-EWS.git](https://github.com/mjharris95/).

-   The performance of each indicator as an early warning signal was then assessed by performing significance tests on the rolling window statistics as follows.

    -   First, **we determined the trend in each indicator over time by calculating Kendall's correlation coefficient**, $\tau$, between the calculated rolling window statistic at each time and the time index *i*.

        -   The sign of indicates the direction of change, while the magnitude corresponds to the strength of the temporal trend.

    -   We used the following permutation approach to assess the significance of the estimated correlations.

        -   We first **generated 10,000 permutations of case incidence during the approach to criticality**, thus preserving the general features of the original data (e.g. amplitudes of fluctuations) while eliminating underlying patterns of serial dependence. We then **performed rolling window analyses on the permuted data to generate a null distribution of correlation coefficients for each indicator. The false positive rates of the 10 tests were estimated by finding the proportion of these permutations with a value for more extreme than that calculated for the Kericho data. The resulting p-values quantify the reliability of their corresponding statistics as early warning signals.** In an applied surveillance setting, time series data would be updated regularly. To reflect this, we first included only data collected at least 96 months prior to the critical transition (April 1985) and repeatedly added an additional month of data, recalculating p-values for each month.

------------------------------------------------------------------------

Okay, so it looks like they just calculated Kendall's $\tau$, and then did some bootstrapping? We won't have to do that second bit for the simulated data, so let's start there.

# Raw, Simulated Data (Sliding Window = 15 Days)

```{r, eval = FALSE}
constant.sims.var.15.tau
constant.sims.mean.15.tau
constant.sims.cv.15.tau
constant.sims.sk.15.tau
constant.sims.kt.15.tau
constant.sims.id.15.tau
constant.sims.ac.15.tau
constant.sims.acv.15.tau
constant.sims.fdv.15.tau
constant.sims.dt.15.tau
warming.sims.var.15.tau
warming.sims.mean.15.tau
warming.sims.cv.15.tau
warming.sims.sk.15.tau
warming.sims.kt.15.tau
warming.sims.id.15.tau
warming.sims.ac.15.tau
warming.sims.acv.15.tau
warming.sims.fdv.15.tau
warming.sims.dt.15.tau
```

Loading some more packages for plotting:

```{r}
library(ggplot2)
library(hrbrthemes)
library(dplyr)
library(tidyr)
library(viridis)
```

## Mean

```{r}
print("Summary of CT KTs:")
summary(constant.sims.mean.15.tau)
print("Summary of WT KTs:")
summary(warming.sims.mean.15.tau)
print("Proportion of WT simulations with larger KT than 0:")
(length(warming.sims.mean.15.tau) - length(which(warming.sims.mean.15.tau < 0)))/length(warming.sims.mean.15.tau)
print("Proportion of CT simulations with a larger KT than the median WT KT:")
length(which(constant.sims.mean.15.tau >= median(warming.sims.mean.15.tau)))/length(constant.sims.mean.15.tau)

raw.sim.mean.15.tau <- data.frame(cbind(constant.sims.mean.15.tau, warming.sims.mean.15.tau))
Treatment <- c(rep("Constant", 1000), rep("Warming", 1000))
k.tau <- c(constant.sims.mean.15.tau, warming.sims.mean.15.tau)
raw.sim.mean.15.tau <- data.frame(
  Treatment = Treatment, 
  k.tau = k.tau
)
saveRDS(raw.sim.mean.15.tau, "Objects/raw.sim.mean.15.tau.rds")

cols <- c(rep("#1434A4", 512), rep("#C21E56", 512))
cols.pt <- c(rep("#A7C7E7", 1000), rep("#F6799D", 1000))

med.const.sims.mean.15.tau <- median(constant.sims.mean.15.tau)
med.warm.sims.mean.15.tau <- median(warming.sims.mean.15.tau)
ggplot(data = raw.sim.mean.15.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.mean.15.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.mean.15.tau, colour = "#C21E56") +
  xlab("Kendall's Tau") +
  ylab("Density (Mean)") +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))

ggplot(raw.sim.mean.15.tau, aes(x = Treatment, y = k.tau)) + 
  geom_violin(fill = cols, colour = cols) +
  geom_jitter(height = 0, width = 0.2, colour = cols.pt) +
  stat_summary(fun.y = median, geom = "point", size = 6, colour = "black", shape = 18) + 
  xlab("Mean") +
  ylab("Kendall's Tau") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, colour = "black"),
        text = element_text(size = 12)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))
```

## Skewness

```{r}
print("Summary of CT KTs:")
summary(constant.sims.sk.15.tau)
print("Summary of WT KTs:")
summary(warming.sims.sk.15.tau)
print("Proportion of WT simulations with larger KT than 0:")
(length(warming.sims.sk.15.tau) - length(which(warming.sims.sk.15.tau < 0)))/length(warming.sims.sk.15.tau)
print("Proportion of CT simulations with a larger KT than the median WT KT:")
length(which(constant.sims.sk.15.tau >= median(warming.sims.sk.15.tau)))/length(constant.sims.sk.15.tau)

raw.sim.sk.15.tau <- data.frame(cbind(constant.sims.sk.15.tau, warming.sims.sk.15.tau))
Treatment <- c(rep("Constant", 1000), rep("Warming", 1000))
k.tau <- c(constant.sims.sk.15.tau, warming.sims.sk.15.tau)
raw.sim.sk.15.tau <- data.frame(
  Treatment = Treatment, 
  k.tau = k.tau
)
saveRDS(raw.sim.sk.15.tau, "Objects/raw.sim.sk.15.tau.rds")

cols <- c(rep("#1434A4", 512), rep("#C21E56", 512))
cols.pt <- c(rep("#A7C7E7", 1000), rep("#F6799D", 1000))

med.const.sims.sk.15.tau <- median(constant.sims.sk.15.tau)
med.warm.sims.sk.15.tau <- median(warming.sims.sk.15.tau)
ggplot(data = raw.sim.sk.15.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.sk.15.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.sk.15.tau, colour = "#C21E56") +
  xlab("Kendall's Tau") +
  ylab("Density (Skewness)") +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))

ggplot(raw.sim.sk.15.tau, aes(x = Treatment, y = k.tau)) + 
  geom_violin(fill = cols, colour = cols) +
  geom_jitter(height = 0, width = 0.2, colour = cols.pt) +
  stat_summary(fun.y = median, geom = "point", size = 6, colour = "black", shape = 18) + 
  xlab("Skewness") +
  ylab("Kendall's Tau") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, colour = "black"),
        text = element_text(size = 12)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))
```

## Kurtosis

```{r}
print("Summary of CT KTs:")
summary(constant.sims.kt.15.tau)
print("Summary of WT KTs:")
summary(warming.sims.kt.15.tau)
print("Proportion of WT simulations with larger KT than 0:")
(length(warming.sims.kt.15.tau) - length(which(warming.sims.kt.15.tau < 0)))/length(warming.sims.kt.15.tau)
print("Proportion of CT simulations with a larger KT than the median WT KT:")
length(which(constant.sims.kt.15.tau >= median(warming.sims.kt.15.tau)))/length(constant.sims.kt.15.tau)

raw.sim.kt.15.tau <- data.frame(cbind(constant.sims.kt.15.tau, warming.sims.kt.15.tau))
Treatment <- c(rep("Constant", 1000), rep("Warming", 1000))
k.tau <- c(constant.sims.kt.15.tau, warming.sims.kt.15.tau)
raw.sim.kt.15.tau <- data.frame(
  Treatment = Treatment, 
  k.tau = k.tau
)
saveRDS(raw.sim.kt.15.tau, "Objects/raw.sim.kt.15.tau.rds")

cols <- c(rep("#1434A4", 512), rep("#C21E56", 512))
cols.pt <- c(rep("#A7C7E7", 1000), rep("#F6799D", 1000))

med.const.sims.kt.15.tau <- median(constant.sims.kt.15.tau)
med.warm.sims.kt.15.tau <- median(warming.sims.kt.15.tau)
ggplot(data = raw.sim.kt.15.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.kt.15.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.kt.15.tau, colour = "#C21E56") +
  xlab("Kendall's Tau") +
  ylab("Density (Kurtosis)") +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))

ggplot(raw.sim.kt.15.tau, aes(x = Treatment, y = k.tau)) + 
  geom_violin(fill = cols, colour = cols) +
  geom_jitter(height = 0, width = 0.2, colour = cols.pt) +
  stat_summary(fun.y = median, geom = "point", size = 6, colour = "black", shape = 18) + 
  xlab("Kurtosis") +
  ylab("Kendall's Tau") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, colour = "black"),
        text = element_text(size = 12)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))
```

## Variance

```{r}
print("Summary of CT KTs:")
summary(constant.sims.var.15.tau)
print("Summary of WT KTs:")
summary(warming.sims.var.15.tau)
print("Proportion of WT simulations with larger KT than 0:")
(length(warming.sims.var.15.tau) - length(which(warming.sims.var.15.tau < 0)))/length(warming.sims.var.15.tau)
print("Proportion of CT simulations with a larger KT than the median WT KT:")
length(which(constant.sims.var.15.tau >= median(warming.sims.var.15.tau)))/length(constant.sims.var.15.tau)

raw.sim.var.15.tau <- data.frame(cbind(constant.sims.var.15.tau, warming.sims.var.15.tau))
Treatment <- c(rep("Constant", 1000), rep("Warming", 1000))
k.tau <- c(constant.sims.var.15.tau, warming.sims.var.15.tau)
raw.sim.var.15.tau <- data.frame(
  Treatment = Treatment, 
  k.tau = k.tau
)
saveRDS(raw.sim.var.15.tau, "Objects/raw.sim.var.15.tau.rds")

cols <- c(rep("#1434A4", 512), rep("#C21E56", 512))
cols.pt <- c(rep("#A7C7E7", 1000), rep("#F6799D", 1000))

med.const.sims.var.15.tau <- median(constant.sims.var.15.tau)
med.warm.sims.var.15.tau <- median(warming.sims.var.15.tau)
ggplot(data = raw.sim.var.15.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.var.15.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.var.15.tau, colour = "#C21E56") +
  xlab("Kendall's Tau") +
  ylab("Density (Variance)") +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))

ggplot(raw.sim.var.15.tau, aes(x = Treatment, y = k.tau)) + 
  geom_violin(fill = cols, colour = cols) +
  geom_jitter(height = 0, width = 0.2, colour = cols.pt) +
  stat_summary(fun.y = median, geom = "point", size = 6, colour = "black", shape = 18) + 
  xlab("Variance") +
  ylab("Kendall's Tau") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, colour = "black"),
        text = element_text(size = 12)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))
```

## Co-Efficient of Variation

```{r}
print("Summary of CT KTs:")
summary(constant.sims.cv.15.tau)
print("Summary of WT KTs:")
summary(warming.sims.cv.15.tau)
print("Proportion of WT simulations with larger KT than 0:")
(length(warming.sims.cv.15.tau) - length(which(warming.sims.cv.15.tau < 0)))/length(warming.sims.cv.15.tau)
print("Proportion of CT simulations with a larger KT than the median WT KT:")
length(which(constant.sims.cv.15.tau >= median(warming.sims.cv.15.tau)))/length(constant.sims.cv.15.tau)

raw.sim.cv.15.tau <- data.frame(cbind(constant.sims.cv.15.tau, warming.sims.cv.15.tau))
Treatment <- c(rep("Constant", 1000), rep("Warming", 1000))
k.tau <- c(constant.sims.cv.15.tau, warming.sims.cv.15.tau)
raw.sim.cv.15.tau <- data.frame(
  Treatment = Treatment, 
  k.tau = k.tau
)
saveRDS(raw.sim.cv.15.tau, "Objects/raw.sim.cv.15.tau.rds")

cols <- c(rep("#1434A4", 512), rep("#C21E56", 512))
cols.pt <- c(rep("#A7C7E7", 1000), rep("#F6799D", 1000))

med.const.sims.cv.15.tau <- median(constant.sims.cv.15.tau)
med.warm.sims.cv.15.tau <- median(warming.sims.cv.15.tau)
ggplot(data = raw.sim.cv.15.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.cv.15.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.cv.15.tau, colour = "#C21E56") +
  xlab("Kendall's Tau") +
  ylab("Density (Co-Efficient of Variation)") +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))

ggplot(raw.sim.cv.15.tau, aes(x = Treatment, y = k.tau)) + 
  geom_violin(fill = cols, colour = cols) +
  geom_jitter(height = 0, width = 0.2, colour = cols.pt) +
  stat_summary(fun.y = median, geom = "point", size = 6, colour = "black", shape = 18) + 
  xlab("Co-Efficient of Variation") +
  ylab("Kendall's Tau") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, colour = "black"),
        text = element_text(size = 12)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))
```

## Index of Dispersion

```{r}
print("Summary of CT KTs:")
summary(constant.sims.id.15.tau)
print("Summary of WT KTs:")
summary(warming.sims.id.15.tau)
print("Proportion of WT simulations with larger KT than 0:")
(length(warming.sims.id.15.tau) - length(which(warming.sims.id.15.tau < 0)))/length(warming.sims.id.15.tau)
print("Proportion of CT simulations with a larger KT than the median WT KT:")
length(which(constant.sims.id.15.tau >= median(warming.sims.id.15.tau)))/length(constant.sims.id.15.tau)

raw.sim.id.15.tau <- data.frame(cbind(constant.sims.id.15.tau, warming.sims.id.15.tau))
Treatment <- c(rep("Constant", 1000), rep("Warming", 1000))
k.tau <- c(constant.sims.id.15.tau, warming.sims.id.15.tau)
raw.sim.id.15.tau <- data.frame(
  Treatment = Treatment, 
  k.tau = k.tau
)
saveRDS(raw.sim.id.15.tau, "Objects/raw.sim.id.15.tau.rds")

cols <- c(rep("#1434A4", 512), rep("#C21E56", 512))
cols.pt <- c(rep("#A7C7E7", 1000), rep("#F6799D", 1000))

med.const.sims.id.15.tau <- median(constant.sims.id.15.tau)
med.warm.sims.id.15.tau <- median(warming.sims.id.15.tau)
ggplot(data = raw.sim.id.15.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.id.15.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.id.15.tau, colour = "#C21E56") +
  xlab("Kendall's Tau") +
  ylab("Density (Index of Dispersion)") +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))

ggplot(raw.sim.id.15.tau, aes(x = Treatment, y = k.tau)) + 
  geom_violin(fill = cols, colour = cols) +
  geom_jitter(height = 0, width = 0.2, colour = cols.pt) +
  stat_summary(fun.y = median, geom = "point", size = 6, colour = "black", shape = 18) + 
  xlab("Index of Dispersion") +
  ylab("Kendall's Tau") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, colour = "black"),
        text = element_text(size = 12)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))
```

## First Difference Variance

```{r}
print("Summary of CT KTs:")
summary(constant.sims.fdv.15.tau)
print("Summary of WT KTs:")
summary(warming.sims.fdv.15.tau)
print("Proportion of WT simulations with larger KT than 0:")
(length(warming.sims.fdv.15.tau) - length(which(warming.sims.fdv.15.tau < 0)))/length(warming.sims.fdv.15.tau)
print("Proportion of CT simulations with a larger KT than the median WT KT:")
length(which(constant.sims.fdv.15.tau >= median(warming.sims.fdv.15.tau)))/length(constant.sims.fdv.15.tau)

raw.sim.fdv.15.tau <- data.frame(cbind(constant.sims.fdv.15.tau, warming.sims.fdv.15.tau))
Treatment <- c(rep("Constant", 1000), rep("Warming", 1000))
k.tau <- c(constant.sims.fdv.15.tau, warming.sims.fdv.15.tau)
raw.sim.fdv.15.tau <- data.frame(
  Treatment = Treatment, 
  k.tau = k.tau
)
saveRDS(raw.sim.fdv.15.tau, "Objects/raw.sim.fdv.15.tau.rds")

cols <- c(rep("#1434A4", 512), rep("#C21E56", 512))
cols.pt <- c(rep("#A7C7E7", 1000), rep("#F6799D", 1000))

med.const.sims.fdv.15.tau <- median(constant.sims.fdv.15.tau)
med.warm.sims.fdv.15.tau <- median(warming.sims.fdv.15.tau)
ggplot(data = raw.sim.fdv.15.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.fdv.15.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.fdv.15.tau, colour = "#C21E56") +
  xlab("Kendall's Tau") +
  ylab("Density (First Difference Variance)") +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))

ggplot(raw.sim.fdv.15.tau, aes(x = Treatment, y = k.tau)) + 
  geom_violin(fill = cols, colour = cols) +
  geom_jitter(height = 0, width = 0.2, colour = cols.pt) +
  stat_summary(fun.y = median, geom = "point", size = 6, colour = "black", shape = 18) + 
  xlab("First Difference Variance") +
  ylab("Kendall's Tau") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, colour = "black"),
        text = element_text(size = 12)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))
```

## Autocorrelation

```{r}
print("Summary of CT KTs:")
summary(constant.sims.ac.15.tau)
print("Summary of WT KTs:")
summary(warming.sims.ac.15.tau)
print("Proportion of WT simulations with larger KT than 0:")
(length(warming.sims.ac.15.tau) - length(which(warming.sims.ac.15.tau < 0)))/length(warming.sims.ac.15.tau)
print("Proportion of CT simulations with a larger KT than the median WT KT:")
length(which(constant.sims.ac.15.tau >= median(warming.sims.ac.15.tau)))/length(constant.sims.ac.15.tau)

raw.sim.ac.15.tau <- data.frame(cbind(constant.sims.ac.15.tau, warming.sims.ac.15.tau))
Treatment <- c(rep("Constant", 1000), rep("Warming", 1000))
k.tau <- c(constant.sims.ac.15.tau, warming.sims.ac.15.tau)
raw.sim.ac.15.tau <- data.frame(
  Treatment = Treatment, 
  k.tau = k.tau
)
saveRDS(raw.sim.ac.15.tau, "Objects/raw.sim.ac.15.tau.rds")

cols <- c(rep("#1434A4", 512), rep("#C21E56", 512))
cols.pt <- c(rep("#A7C7E7", 1000), rep("#F6799D", 1000))

med.const.sims.ac.15.tau <- median(constant.sims.ac.15.tau)
med.warm.sims.ac.15.tau <- median(warming.sims.ac.15.tau)
ggplot(data = raw.sim.ac.15.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.ac.15.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.ac.15.tau, colour = "#C21E56") +
  xlab("Kendall's Tau") +
  ylab("Density (Autocorrelation)") +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))

ggplot(raw.sim.ac.15.tau, aes(x = Treatment, y = k.tau)) + 
  geom_violin(fill = cols, colour = cols) +
  geom_jitter(height = 0, width = 0.2, colour = cols.pt) +
  stat_summary(fun.y = median, geom = "point", size = 6, colour = "black", shape = 18) + 
  xlab("Autocorrelation") +
  ylab("Kendall's Tau") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, colour = "black"),
        text = element_text(size = 12)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))
```

## Autocovariance

```{r}
print("Summary of CT KTs:")
summary(constant.sims.acv.15.tau)
print("Summary of WT KTs:")
summary(warming.sims.acv.15.tau)
print("Proportion of WT simulations with larger KT than 0:")
(length(warming.sims.acv.15.tau) - length(which(warming.sims.acv.15.tau < 0)))/length(warming.sims.acv.15.tau)
print("Proportion of CT simulations with a larger KT than the median WT KT:")
length(which(constant.sims.acv.15.tau >= median(warming.sims.acv.15.tau)))/length(constant.sims.acv.15.tau)

raw.sim.acv.15.tau <- data.frame(cbind(constant.sims.acv.15.tau, warming.sims.acv.15.tau))
Treatment <- c(rep("Constant", 1000), rep("Warming", 1000))
k.tau <- c(constant.sims.acv.15.tau, warming.sims.acv.15.tau)
raw.sim.acv.15.tau <- data.frame(
  Treatment = Treatment, 
  k.tau = k.tau
)
saveRDS(raw.sim.acv.15.tau, "Objects/raw.sim.acv.15.tau.rds")

cols <- c(rep("#1434A4", 512), rep("#C21E56", 512))
cols.pt <- c(rep("#A7C7E7", 1000), rep("#F6799D", 1000))

med.const.sims.acv.15.tau <- median(constant.sims.acv.15.tau)
med.warm.sims.acv.15.tau <- median(warming.sims.acv.15.tau)
ggplot(data = raw.sim.acv.15.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.acv.15.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.acv.15.tau, colour = "#C21E56") +
  xlab("Kendall's Tau") +
  ylab("Density (Autocovariance)") +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))

ggplot(raw.sim.acv.15.tau, aes(x = Treatment, y = k.tau)) + 
  geom_violin(fill = cols, colour = cols) +
  geom_jitter(height = 0, width = 0.2, colour = cols.pt) +
  stat_summary(fun.y = median, geom = "point", size = 6, colour = "black", shape = 18) + 
  xlab("Autocovariance") +
  ylab("Kendall's Tau") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, colour = "black"),
        text = element_text(size = 12)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))
```

## Decay Time

```{r}
constant.sims.dt.15.tau.rmna <- constant.sims.dt.15.tau[!is.na(constant.sims.dt.15.tau)]
warming.sims.dt.15.tau.rmna <- warming.sims.dt.15.tau[!is.na(warming.sims.dt.15.tau)]

print("Summary of CT KTs:")
summary(constant.sims.dt.15.tau.rmna)
print("Summary of WT KTs:")
summary(warming.sims.dt.15.tau.rmna)
print("Proportion of WT simulations with larger KT than 0:")
(length(warming.sims.dt.15.tau.rmna) - length(which(warming.sims.dt.15.tau.rmna < 0)))/length(warming.sims.dt.15.tau.rmna)
print("Proportion of CT simulations with a larger KT than the median WT KT:")
length(which(constant.sims.dt.15.tau.rmna >= median(warming.sims.dt.15.tau.rmna)))/length(constant.sims.dt.15.tau.rmna)

raw.sim.dt.15.tau <- data.frame(cbind(constant.sims.dt.15.tau, warming.sims.dt.15.tau))
Treatment <- c(rep("Constant", 1000), rep("Warming", 1000))
k.tau <- c(constant.sims.dt.15.tau, warming.sims.dt.15.tau)
raw.sim.dt.15.tau <- data.frame(
  Treatment = Treatment, 
  k.tau = k.tau
)
saveRDS(raw.sim.dt.15.tau, "Objects/raw.sim.dt.15.tau.rds")

cols <- c(rep("#1434A4", 512), rep("#C21E56", 512))
cols.pt <- c(rep("#A7C7E7", 1000), rep("#F6799D", 1000))

med.const.sims.dt.15.tau <- median(constant.sims.dt.15.tau.rmna)
med.warm.sims.dt.15.tau <- median(warming.sims.dt.15.tau.rmna)
ggplot(data = raw.sim.dt.15.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.dt.15.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.dt.15.tau, colour = "#C21E56") +
  xlab("Kendall's Tau") +
  ylab("Density (Decay Time)") +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))

ggplot(raw.sim.dt.15.tau, aes(x = Treatment, y = k.tau)) + 
  geom_violin(fill = cols, colour = cols) +
  geom_jitter(height = 0, width = 0.2, colour = cols.pt) +
  stat_summary(fun.y = median, geom = "point", size = 6, colour = "black", shape = 18) + 
  xlab("Decay Time") +
  ylab("Kendall's Tau") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, colour = "black"),
        text = element_text(size = 12)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))
```

## Plotting

```{r}
df <- data.frame(
  metric = c(rep("Mean", 2000),
             rep("Skewness", 2000),
             rep("Kurtosis", 2000),
             rep("Variance", 2000),
             rep("Co-Efficient of Variance", 2000),
             rep("Index of Dispersion", 2000),
             rep("First Difference Variance", 2000),
             rep("Autocorrelation", 2000),
             rep("Autocovariance", 2000),
             rep("Decay Time", 2000)),
  Treatment = c(rep(c(rep("Constant", 1000), rep("Warming", 1000)), 10)),
  k.tau = c(constant.sims.mean.15.tau, warming.sims.mean.15.tau,
            constant.sims.sk.15.tau, warming.sims.sk.15.tau,
            constant.sims.kt.15.tau, warming.sims.kt.15.tau,
            constant.sims.var.15.tau, warming.sims.var.15.tau,
            constant.sims.cv.15.tau, warming.sims.cv.15.tau,
            constant.sims.id.15.tau, warming.sims.id.15.tau,
            constant.sims.fdv.15.tau, warming.sims.fdv.15.tau,
            constant.sims.ac.15.tau, warming.sims.ac.15.tau,
            constant.sims.acv.15.tau, warming.sims.acv.15.tau,
            constant.sims.dt.15.tau, warming.sims.dt.15.tau)
)

df$metric <- factor(df$metric, levels = unique(df$metric))
cols <- rep(c("#1434A4","#C21E56"), 10)
cols.pts <- rep(c(rep("#1434A4", 1000), rep("#C21E56", 1000)), 10)

df %>%
  ggplot(aes(x = metric, y = k.tau, fill = Treatment)) +
  geom_violin(position = position_dodge(width = 1), alpha = 0.2) +
  scale_fill_manual(values = cols) +
  # geom_jitter(height = 0, width = 0, colour = "black", alpha = 0.5) +
  geom_point(aes(y = k.tau), 
             position = position_jitterdodge(seed = 1, dodge.width = 1), 
             colour = cols.pts, size = 0.25, shape = 8) + 
  stat_summary(fun.y = median, geom = "point", position = position_dodge(width = 1), size = 4, colour = "black", shape = 18) +
  xlab("") +
  ylab("Kendall's Tau") +
  ylim(-1, 1) +
  geom_hline(yintercept = 0, linetype = "dotted", color = "black") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, colour = "black"),
        text = element_text(size = 12)) +
  theme(axis.text.y = element_text(colour = "black"),
        text = element_text(size = 12)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))
```

# Raw, Empirical Data

`constant.mean.tau` and `warming.mean.tau`, etc., taken from `4_EWS-EmpData.Rmd`.

```{r}
emp.ktaus <- data.frame(
  mean = c(constant.mean.tau, warming.mean.tau),
  sk = c(constant.sk.tau, warming.sk.tau),
  kt = c(constant.kt.tau, warming.kt.tau),
  var = c(constant.var.tau, warming.var.tau),
  cv = c(constant.cv.tau, warming.cv.tau),
  id = c(constant.id.tau, warming.id.tau),
  fdv = c(constant.fdv.tau, warming.fdv.tau),
  ac = c(constant.ac.tau, warming.ac.tau),
  acv = c(constant.acv.tau, warming.acv.tau),
  dt = c(constant.dt.tau, warming.dt.tau)
)
rownames(emp.ktaus) <- c("constant.1", "constant.2", "constant.3", "constant.4", "warming.1", "warming.2", "warming.3", "warming.4")

x.labs <- c("Mean", "Skewness", "Kurtosis", 
                              "Variance", "Co-Efficient of Variation", "Index of Dispersion",
                              "First Difference Variance", "Autocorrelation", "Autocovariance",
                              "Decay Time")
par(mar = c(6.5, 4, 4, 4)) # Order: 'bottom', 'left', 'top', 'right'.
plot(x = 1, type = "n", xaxt = "n", yaxt = "n", 
     main = "Experimental Data", xlab = "", ylab = "Kendall's Tau",
     xlim = c(1, 10), ylim = c(-1, 1))
text(1:10, emp.ktaus[1, ], label = 1, col = alpha("#1434A4", 0.5))
text(1:10, emp.ktaus[2, ], label = 2, col = alpha("#1434A4", 0.5))
text(1:10, emp.ktaus[3, ], label = 3, col = alpha("#1434A4", 0.5))
text(1:10, emp.ktaus[4, ], label = 4, col = alpha("#1434A4", 0.5))
text(1:10, emp.ktaus[5, ], label = 5, col = alpha("#C21E56", 0.5))
text(1:10, emp.ktaus[6, ], label = 6, col = alpha("#C21E56", 0.5))
text(1:10, emp.ktaus[7, ], label = 7, col = alpha("#C21E56", 0.5))
text(1:10, emp.ktaus[8, ], label = 8, col = alpha("#C21E56", 0.5))
abline(h = 0, lty = 3)
# axis(1, at = 1:10, labels = c("Mean", "Skewness", "Kurtosis", 
#                               "Variance", "Co-Efficient of Variation", "Index of Dispersion",
#                               "First Difference Variance", "Autocorrelation", "Autocovariance",
#                               "Decay Time"))
axis(2, at = c(-1, -0.5, 0, 0.5, 1))
text(cex = 1, x = 1:10, y = -1.4, x.labs, xpd = TRUE, srt = 35)
legend(8, 1.1, legend = c("Constant Tempertaure", "Rising Temperature"),
       pch = 19, col = c("#1434A4", "#C21E56"), cex = 0.8, bty = "n")

df <- data.frame(
  metric = c(rep("Mean", 8),
             rep("Skewness", 8),
             rep("Kurtosis", 8),
             rep("Variance", 8),
             rep("Co-Efficient of Variance", 8),
             rep("Index of Dispersion", 8),
             rep("First Difference Variance", 8),
             rep("Autocorrelation", 8),
             rep("Autocovariance", 8),
             rep("Decay Time", 8)),
  Treatment = c(rep(c(rep("Constant", 4), rep("Warming", 4)), 10)),
  k.tau = c(constant.mean.tau, warming.mean.tau,
            constant.sk.tau, warming.sk.tau,
            constant.kt.tau, warming.kt.tau,
            constant.var.tau, warming.var.tau,
            constant.cv.tau, warming.cv.tau,
            constant.id.tau, warming.id.tau,
            constant.fdv.tau, warming.fdv.tau,
            constant.ac.tau, warming.ac.tau,
            constant.acv.tau, warming.acv.tau,
            constant.dt.tau, warming.dt.tau)
)

df$metric <- factor(df$metric, levels = unique(df$metric))
cols <- rep(c("#1434A4","#C21E56"), 10)
cols.pts <- rep(c(rep("#1434A4", 4), rep("#C21E56", 4)), 10)

df %>%
  ggplot(aes(x = metric, y = k.tau, fill = Treatment)) +
  geom_violin(position = position_dodge(width = 1), alpha = 0.2) +
  scale_fill_manual(values = cols) +
  # geom_jitter(height = 0, width = 0, colour = "black", alpha = 0.5) +
  geom_point(aes(y = k.tau), 
             position = position_jitterdodge(seed = 1, dodge.width = 1), 
             colour = cols.pts, size = 2, shape = 8) + 
  xlab("") +
  ylab("Kendall's Tau") +
  ylim(-1, 1) +
  geom_hline(yintercept = 0, linetype = "dotted", color = "black") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, colour = "black"),
        text = element_text(size = 12)) +
  theme(axis.text.y = element_text(colour = "black"),
        text = element_text(size = 12)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"))
```
